core: Extract out .getViewSchema() from permission filter
Change-type: patch
