+++ lib/core/permission-filter.js
@@ -7,6 +7,7 @@
 const _ = require('lodash')
 const jsone = require('json-e')
 const jsonSchema = require('./json-schema')
+const views = require('./views')
 const errors = require('./errors')
 const CARDS = require('./cards')
 const assert = require('../assert')
@@ -192,39 +193,6 @@ exports.getSessionUser = async (context, backend, session) => {
 }
 
 /**
- * @summary Get the schema of a view card
- * @function
- * @private
- *
- * @param {Object} card - view card
- * @returns {(Object|Null)} schema
- */
-exports.getViewSchema = (card) => {
-	if (!card) {
-		return null
-	}
-
-	if (card.data && card.data.schema) {
-		return card.data.schema
-	}
-
-	const conjunctions = _.map(_.get(card, [ 'data', 'allOf' ]), 'schema')
-	const disjunctions = _.map(_.get(card, [ 'data', 'anyOf' ]), 'schema')
-
-	if (_.isEmpty(conjunctions) && _.isEmpty(disjunctions)) {
-		return null
-	}
-
-	if (!_.isEmpty(disjunctions)) {
-		conjunctions.push({
-			anyOf: disjunctions
-		})
-	}
-
-	return jsonSchema.merge(conjunctions)
-}
-
-/**
  * @summary Get the filter schemas for the user's roles
  * @function
  * @private
@@ -360,7 +328,7 @@ exports.getQuery = async (context, backend, session, schema) => {
 	}
 
 	const finalSchema = schema.type === CARDS.view.slug
-		? exports.getViewSchema(schema)
+		? views.getSchema(schema)
 		: schema
 	const evaledSchema = evalSchema(finalSchema, {
 		user
